chore: Project refactorement
