+++ src/video_downloader/constants.py
@@ -1,0 +1,17 @@
+from enum import Enum
+import threading
+
+SENTINEL = "STOP_WORKER"
+stop_workers = threading.Event()
+
+
+class PROCESS_STATUS(Enum):
+    SUCCESS = 1
+    CANCELLED = 2
+    FAILED = 3
+
+
+class CONTENT_MODE(Enum):
+    VIDEO = "video"
+    AUDIO = "audio"
+    STREAM = "stream"
